refactor: cleanup view switcher and filter constants
